Fixed the potential integer overflow.

+++ src/gofs/file.go
@@ -55,11 +55,10 @@ func (file *DataFile) Read(p []byte) (n int, err error) {
 }
 
 func (file *DataFile) Write(p []byte) (n int, err error) {
-  // TODO: A really bad idea. Would overflow with seeks > 2^31 - 1, IE, > 2GB
-  needed := int(file.seek) + len(p)
+  needed := file.seek + int64(len(p))
   fmt.Println("Needed:", needed, "Have:", cap(file.data))
 
-  if needed - cap(file.data) > 0 {
+  if needed - int64(cap(file.data)) > 0 {
     newData := make([]byte, needed, needed * 2)
     copy(file.data, newData)
     file.data = newData
